better datacenter endpoints

+++ lib/Conch/Controller/DatacenterRoom.pm
@@ -5,6 +5,10 @@ use Mojo::Base 'Mojolicious::Controller', -signatures;
 use Role::Tiny::With;
 with 'Conch::Role::MojoLog';
 
+use Conch::UUID 'is_uuid';
+
+=pod
+
 =head1 NAME
 
 Conch::Controller::DatacenterRoom
@@ -13,103 +17,97 @@ Conch::Controller::DatacenterRoom
 
 =head2 find_datacenter_room
 
-Handles looking up the object by id or name depending on the url pattern
+Handles looking up the object by id.
 
 =cut
 
 sub find_datacenter_room ($c) {
-	unless($c->is_system_admin) {
-		$c->status(403);
-		return undef;
-	}
+    return $c->status(403) if not $c->is_system_admin;
 
-	if ($c->stash('datacenter_room_id_or_name') =~ /^(.+?)\=(.+)$/) {
-		$c->log->warn("Unsupported identifier '$1'");
-		return $c->status(501);
-	}
+    my $room_id = $c->stash('datacenter_room_id');
+    if (not is_uuid($room_id)) {
+        $c->log->warn('input failed validation');
+        return $c->status(400 => { error => "Datacenter Room ID must be a UUID. Got '$room_id'." });
+    }
 
-	$c->log->debug("Looking up datacenter room ".$c->stash('datacenter_room_id_or_name'));
-	my $room = $c->db_datacenter_rooms->find($c->stash('datacenter_room_id_or_name'));
+    $c->log->debug('Looking up datacenter room '.$room_id);
+    my $room = $c->db_datacenter_rooms->find($room_id);
 
-	if (not $room) {
-		$c->log->debug("Could not find datacenter room");
-		return $c->status(404 => { error => "Not found" });
-	}
+    if (not $room) {
+        $c->log->debug('Could not find datacenter room');
+        return $c->status(404 => { error => 'Not found' });
+    }
 
-	$c->log->debug("Found datacenter room");
-	$c->stash('datacenter_room' => $room);
-	return 1;
+    $c->log->debug('Found datacenter room');
+    $c->stash('datacenter_room' => $room);
+    return 1;
 }
-
 
 =head2 get_all
 
-Get all datacenter rooms
+Get all datacenter rooms.
 
 Response uses the DatacenterRoomsDetailed json schema.
 
 =cut
 
 sub get_all ($c) {
-	return $c->status(403) unless $c->is_system_admin;
+    return $c->status(403) unless $c->is_system_admin;
 
-	my @rooms = $c->db_datacenter_rooms->all;
-	$c->log->debug('Found ' . scalar(@rooms) . ' datacenter rooms');
+    my @rooms = $c->db_datacenter_rooms->all;
+    $c->log->debug('Found ' . scalar(@rooms) . ' datacenter rooms');
 
-	return $c->status(200 => \@rooms);
+    return $c->status(200 => \@rooms);
 }
-
 
 =head2 get_one
 
-Get a single datacenter room
+Get a single datacenter room.
 
 Response uses the DatacenterRoomDetailed json schema.
 
 =cut
 
 sub get_one ($c) {
-	return $c->status(403) unless $c->is_system_admin;
-	$c->status(200, $c->stash('datacenter_room'));
+    return $c->status(403) unless $c->is_system_admin;
+    $c->status(200, $c->stash('datacenter_room'));
 }
 
 =head2 create
 
-Create a new datacenter room
+Create a new datacenter room.
 
 =cut
 
 sub create ($c) {
-	return $c->status(403) unless $c->is_system_admin;
-	my $input = $c->validate_input('DatacenterRoomCreate');
-	return if not $input;
+    return $c->status(403) unless $c->is_system_admin;
+    my $input = $c->validate_input('DatacenterRoomCreate');
+    return if not $input;
 
-	$input->{datacenter_id} = delete $input->{datacenter} if exists $input->{datacenter};
+    $input->{datacenter_id} = delete $input->{datacenter} if exists $input->{datacenter};
 
-	my $room = $c->db_datacenter_rooms->create($input);
-	$c->log->debug("Created datacenter room ".$room->id);
-	$c->status(303 => '/room/'.$room->id);
+    my $room = $c->db_datacenter_rooms->create($input);
+    $c->log->debug('Created datacenter room '.$room->id);
+    $c->status(303 => '/room/'.$room->id);
 }
-
 
 =head2 update
 
-Update an existing room
+Update an existing room.
 
 =cut
 
 sub update ($c) {
-	return $c->status(403) unless $c->is_system_admin;
-	my $input = $c->validate_input('DatacenterRoomUpdate');
-	return if not $input;
+    return $c->status(403) unless $c->is_system_admin;
+    my $input = $c->validate_input('DatacenterRoomUpdate');
+    return if not $input;
 
-	$input->{datacenter_id} = delete $input->{datacenter} if exists $input->{datacenter};
+    $input->{datacenter_id} = delete $input->{datacenter} if exists $input->{datacenter};
 
-	$c->stash('datacenter_room')->update({ %$input, updated => \'NOW()' });
-	$c->log->debug("Updated datacenter room ".$c->stash('datacenter_room_id_or_name'));
-	$c->status(303 => "/room/".$c->stash('datacenter_room')->id);
+    $c->stash('datacenter_room')->update({ %$input, updated => \'now()' });
+    $c->log->debug('Updated datacenter room '.$c->stash('datacenter_room_id'));
+    $c->status(303 => '/room/'.$c->stash('datacenter_room')->id);
 }
-
 
 =head2 delete
 
@@ -120,16 +118,20 @@ Also removes the room from all workspaces.
 =cut
 
 sub delete ($c) {
-	return $c->status(403) unless $c->is_system_admin;
-	# FIXME: if we have cascade_copy => 1 set on this rel,
-	# then we don't have to do this... and we don't have to worry about rack updates either.
-	# But for now, we have a dangling reference to the deleted room in datacenter_rack!
-	$c->stash('datacenter_room')->delete_related('workspace_datacenter_rooms');
-	$c->stash('datacenter_room')->delete;
-	$c->log->debug("Deleted datacenter room ".$c->stash('datacenter_room')->id);
-	return $c->status(204);
-}
+    if ($c->stash('datacenter_room')->related_resultset('datacenter_racks')->exists) {
+        $c->log->debug('Cannot delete datacenter_room: in use by one or more datacenter_racks');
+        return $c->status(400 => { error => 'cannot delete a datacenter_room when a detacenter_rack is referencing it' });
+    }
 
+    return $c->status(403) unless $c->is_system_admin;
+    # FIXME: if we have cascade_copy => 1 set on this rel,
+    # then we don't have to do this... and we don't have to worry about rack updates either.
+    # But for now, we have a dangling reference to the deleted room in datacenter_rack!
+    $c->stash('datacenter_room')->delete_related('workspace_datacenter_rooms');
+    $c->stash('datacenter_room')->delete;
+    $c->log->debug('Deleted datacenter room '.$c->stash('datacenter_room')->id);
+    return $c->status(204);
+}
 
 =head2 racks
 
@@ -138,17 +140,12 @@ Response uses the Racks json schema.
 =cut
 
 sub racks ($c) {
-	return $c->status(403) unless $c->is_system_admin;
+    return $c->status(403) unless $c->is_system_admin;
 
-	my @racks = $c->db_datacenter_racks->search({ datacenter_room_id => $c->stash('datacenter_room')->id });
-	$c->log->debug(
-		"Found ".scalar(@racks).
-		" racks for datacenter room ".$c->stash('datacenter_room')->id
-	);
-	return $c->status(200 => \@racks);
-
+    my @racks = $c->stash('datacenter_room')->related_resultset('datacenter_racks')->all;
+    $c->log->debug('Found '.scalar(@racks).' racks for datacenter room '.$c->stash('datacenter_room')->id);
+    return $c->status(200 => \@racks);
 }
-
 
 1;
 __END__
@@ -164,3 +161,6 @@ v.2.0. If a copy of the MPL was not distributed with this file, You can obtain
 one at http://mozilla.org/MPL/2.0/.
 
 =cut
+
+1;
+# vim: set ts=4 sts=4 sw=4 et :
